feat: Implement shared authentication form for login and registration pages.
